refactor: unify disablePriceImpact conditions
