+++ apps/cowswap-frontend/src/modules/limitOrders/containers/TradeButtons/limitOrdersTradeButtonsMap.tsx
@@ -23,7 +23,4 @@ export const limitOrdersTradeButtonsMap: { [key in LimitOrdersFormState]: Button
   [LimitOrdersFormState.ZeroPrice]: {
     text: 'Invalid price. Try increasing input/output amount.',
   },
-  [LimitOrdersFormState.FeeExceedsFrom]: {
-    text: 'Sell amount is too small',
-  },
 }
