+++ apps/cowswap-frontend/src/modules/limitOrders/hooks/useLimitOrdersFormState.ts
@@ -4,7 +4,6 @@ import { isFractionFalsy } from '@cowprotocol/common-utils'
 import { Currency, CurrencyAmount, Fraction } from '@uniswap/sdk-core'
 
 import { limitRateAtom } from 'modules/limitOrders/state/limitRateAtom'
-import { useTradeQuote } from 'modules/tradeQuote'
 
 import { useSafeMemo } from 'common/hooks/useSafeMemo'
 
@@ -12,7 +11,6 @@ import { useLimitOrdersDerivedState } from './useLimitOrdersDerivedState'
 
 export enum LimitOrdersFormState {
   RateLoading = 'RateLoading',
-  FeeExceedsFrom = 'FeeExceedsFrom',
   PriceIsNotSet = 'PriceIsNotSet',
   ZeroPrice = 'ZeroPrice',
 }
@@ -20,13 +18,12 @@ export enum LimitOrdersFormState {
 interface LimitOrdersFormParams {
   activeRate: Fraction | null
   isRateLoading: boolean
-  feeAmount: CurrencyAmount<Currency> | null
   sellAmount: CurrencyAmount<Currency> | null
   buyAmount: CurrencyAmount<Currency> | null
 }
 
 function getLimitOrdersFormState(params: LimitOrdersFormParams): LimitOrdersFormState | null {
-  const { activeRate, isRateLoading, feeAmount, sellAmount, buyAmount } = params
+  const { activeRate, isRateLoading, sellAmount, buyAmount } = params
 
   if (isFractionFalsy(activeRate)) {
     return LimitOrdersFormState.PriceIsNotSet
@@ -43,27 +40,15 @@ function getLimitOrdersFormState(params: LimitOrdersFormParams): LimitOrdersForm
     return LimitOrdersFormState.ZeroPrice
   }
 
-  if (sellAmount && feeAmount?.greaterThan(sellAmount)) {
-    return LimitOrdersFormState.FeeExceedsFrom
-  }
-
   return null
 }
 
 export function useLimitOrdersFormState(): LimitOrdersFormState | null {
   const { inputCurrencyAmount, outputCurrencyAmount } = useLimitOrdersDerivedState()
-  const quote = useTradeQuote()
   const { activeRate, isLoading } = useAtomValue(limitRateAtom)
-
-  const feeRawAmount = quote?.response?.quote?.feeAmount
-  const feeAmount =
-    feeRawAmount && inputCurrencyAmount
-      ? CurrencyAmount.fromRawAmount(inputCurrencyAmount.currency, feeRawAmount)
-      : null
 
   const params: LimitOrdersFormParams = {
     activeRate,
-    feeAmount,
     isRateLoading: isLoading,
     sellAmount: inputCurrencyAmount,
     buyAmount: outputCurrencyAmount,
